feat(storage): implemented more effective pin logs for blockwise storage
[no changelog]
