+++ storage/tests/python/src/consts.py
@@ -78,23 +78,6 @@ COUNTER_MAX_TAIL = 64
 # Storage key of the PIN entry log and PIN success log.
 PIN_LOG_KEY = (PIN_APP_ID << 8) | 0x01
 
-# Length of items in the PIN entry log
-PIN_LOG_GUARD_KEY_SIZE = 4
-
-# Values used for the guard key integrity check.
-GUARD_KEY_MODULUS = 6311
-GUARD_KEY_REMAINDER = 15
-GUARD_KEY_RANDOM_MAX = (0xFFFFFFFF // GUARD_KEY_MODULUS) + 1
-
-# Length of both success log and entry log
-PIN_LOG_SIZE = 64
-
-# Used for in guard bits operations.
-LOW_MASK = 0x55555555
-
-# Log initialized to all FFs.
-ALL_FF_LOG = 0xFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFF
-
 # ----- Bytes -----
 
 # If the top bit of APP is set, then the value is not encrypted.
